salary variable fix

+++ src/main/java/com/hico/hico/entity/Employee.java
@@ -93,7 +93,7 @@ public class Employee {
         return salary;
     }
 
-    public void setSalary(float salary) {
+    public void setSalary(double salary) {
         this.salary = salary;
     }
